🧪 7N0FHQ code: split pre-merge closure coverage
